added some stuff to the readme, merged the python stuff
